requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pyyaml==6.0.1
httpx==0.27.0
